[PCX-2554] Extract colors and fonts to separate files
