+++ Sources/PayoneerCheckout/Core/Assets/Colors.swift
@@ -1,0 +1,16 @@
+// Copyright (c) 2022 Payoneer Germany GmbH
+// https://www.payoneer.com
+//
+// This file is open source and available under the MIT license.
+// See the LICENSE file for more information.
+
+import UIKit
+
+enum Colors {
+    static let primaryText = UIColor(named: "PrimaryText", in: .current, compatibleWith: nil)!
+    static let secondaryText = UIColor(named: "SecondaryText", in: .current, compatibleWith: nil)!
+    static let background = UIColor(named: "Background", in: .current, compatibleWith: nil)!
+    static let accent = UIColor(named: "Accent", in: .current, compatibleWith: nil)!
+    static let error = UIColor(named: "Error", in: .current, compatibleWith: nil)!
+    static let border = UIColor(named: "Border", in: .current, compatibleWith: nil)!
+}
